Added hierarchical configuration, to enable different settings for docker images and local dev
RECAP: we now have a config folder, with `base.yaml` which contains the default settings for both Docker and local, and two more .yaml files for the settings for the two mentioned environments. The `get_configuration()` function has been edited in order to return the settings based on the `APP_ENVIRONMENT` env variable, defaulting to local if it isn't set.