Using env variables to inject secrets to connection string
RECAP: The connection string might contain values we might not want to commit to version control. With environment variables we can inject secrets at runtime into the app. Now we can customize values in our `Settings` struct, overriding what is specified in our configuration files. This is convenient as it makes it easy to inject values that are too dynamic or too sensitive, making it also faster to change the behavior of our application.

Example: APP_APPLICATION__PORT=5001 would set `Settings.application.port` to 5001